Refactor: 밤 리팩토링 중
- 캐릭터 관련 로직  분리
- 캐릭터 특성 재확인
- 특성 시트 오류 수정
- 특성 최적화 및 구현 수정
- 특성 모두 정상 작동 확인

+++ Assets/Scenes/Night/Script/Class/Item/ChargingReaper.cs
@@ -9,7 +9,7 @@ public class ChargingReaper : MonoBehaviour
     GameObject reaperAfterImage;
     public bool isAttack = false;
     public int chargingGauge = 100;
-    public double reaperAttackDamaege; //임시
+    public float reaperAttackDamaege; //임시
     int itemRank;
 
     public void SetItemRank(int getRank)
