+++ Assets/Scripts/Trait.cs
@@ -1,7 +1,5 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using UnityEngine;
 
 
 [Serializable]
@@ -21,15 +19,15 @@ public class Trait
     public int imageIndex;
     public string script;
     public int effectType1;
-    public double effectValue1;
+    public float effectValue1;
     public bool effectMulti1;
     public int effectType2;
-    public double effectValue2;
+    public float effectValue2;
     public bool effectMulti2;
     public int effectType3;
-    public double effectValue3;
+    public float effectValue3;
     public bool effectMulti3;
     public int effectType4;
-    public double effectValue4;
+    public float effectValue4;
     public bool effectMulti4;
 }
